+++ cogs/commands/time_command.py
@@ -16,11 +16,11 @@ class TimeCommand(discord.Cog):
     async def time(self, ctx: discord.ApplicationContext):
         self.logger.info('Getting time and sending to message')
         utc_time = datetime.datetime.now()
-        mldkyt_time = utc_time + datetime.timedelta(hours=2)
+        programmer_astolfo_time = utc_time + datetime.timedelta(hours=2)
         lang = get_user_lang(ctx.author.id)
         await ctx.respond(get_string('time', lang) % (
             utc_time.strftime("%H:%M"), 
             utc_time.strftime("%I:%M %p"), 
-            mldkyt_time.strftime("%H:%M"), 
-            mldkyt_time.strftime("%I:%M %p")
+            programmer_astolfo_time.strftime("%H:%M"), 
+            programmer_astolfo_time.strftime("%I:%M %p")
             ))