+++ main.py
@@ -40,6 +40,7 @@ from cogs.verification import Verification, VerifyMain
 from cogs.channel_specific.daily_fun_fact_limit import DailyFunFactLimit
 from cogs.commands.language_commands import LanguageCommands
 from views.roles import MainView
+from cogs.commands.brimo_command import BrimoCommand
 
 sentry_sdk.init(
     dsn=constants.sentry_url,
@@ -127,6 +128,8 @@ main_logger.info('Loading module: Femboy of the month')
 bot.add_cog(FemboyOfTheMonth(bot))
 main_logger.info('Loading module: Language Commands')
 bot.add_cog(LanguageCommands(bot))
+main_logger.info('Loading module: Brimo command')
+bot.add_cog(BrimoCommand(bot))
 
 @bot.event
 async def on_ready():
@@ -136,6 +139,16 @@ async def on_ready():
     bot.add_view(VerifyMain())
     bot.add_view(AnalyticsFrom())
 
+# Handle slash command errors, if it's a cooldown error, send a message to the user
+@bot.event
+async def on_slash_command_error(ctx: discord.ApplicationContext, error):
+    if isinstance(error, CommandOnCooldown):
+        embed = Embed(title='You are on cooldown cutie :3',
+                      description=f'Try again in about {int(round(error.retry_after, 0))} seconds :3',
+                      color=Color.red())
+        await ctx.respond(embed=embed, ephemeral=True)
+    else:
+        raise error
 
 @bot.slash_command(guild_ids=[constants.guild_id])
 async def version(ctx: discord.ApplicationContext):