+++ changelog.md
@@ -1,3 +1,4 @@
-# Version 2.4
-- Added languages
-- Added Czech as language
+# Version 2.5
+- Added /brimo command
+- Better language support (If a translation doesn't exist, it will use English)
+- Replaced references to "mldkyt" with "Programmer Astolfo"